+++ seb-esengine/src/no/mnemonic/act/platform/seb/esengine/v1/consumers/FactConsumer.java
@@ -12,7 +12,8 @@ import no.mnemonic.services.common.hazelcast.consumer.TransactionalConsumer;
 import javax.inject.Inject;
 import java.util.Collection;
 
-import static no.mnemonic.act.platform.dao.elastic.FactSearchManager.TargetIndex.*;
+import static no.mnemonic.act.platform.dao.elastic.FactSearchManager.TargetIndex.Daily;
+import static no.mnemonic.act.platform.dao.elastic.FactSearchManager.TargetIndex.TimeGlobal;
 
 /**
  * Component which consumes {@link FactSEB} models, converts them to {@link FactDocument}, and indexes them into ElasticSearch.
@@ -43,9 +44,6 @@ public class FactConsumer implements TransactionalConsumer<FactSEB> {
         } else {
           factSearchManager.indexFact(document, Daily);
         }
-
-        // Compatibility: For now always index into the legacy 'act' index in addition to time global or daily indices.
-        factSearchManager.indexFact(document, Legacy);
       }
     }
   }