+++ service/src/no/mnemonic/act/platform/service/ti/resolvers/IndexSelectCriteriaResolver.java
@@ -1,14 +1,9 @@
 package no.mnemonic.act.platform.service.ti.resolvers;
 
-import no.mnemonic.act.platform.api.exceptions.AccessDeniedException;
-import no.mnemonic.act.platform.api.exceptions.AuthenticationFailedException;
 import no.mnemonic.act.platform.api.exceptions.InvalidArgumentException;
 import no.mnemonic.act.platform.dao.api.criteria.IndexSelectCriteria;
-import no.mnemonic.act.platform.service.contexts.SecurityContext;
-import no.mnemonic.act.platform.service.ti.TiFunctionConstants;
 import no.mnemonic.commons.utilities.ObjectUtils;
 
-import javax.inject.Inject;
 import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
@@ -20,27 +15,17 @@ import java.time.temporal.ChronoUnit;
  */
 public class IndexSelectCriteriaResolver {
 
-  private final SecurityContext securityContext;
-
   static final int MAXIMUM_INDEX_RETENTION_DAYS = 1095; // Must be the same as the ILM policy configured in ElasticSearch.
   private Clock clock = Clock.systemUTC();
 
-  @Inject
-  public IndexSelectCriteriaResolver(SecurityContext securityContext) {
-    this.securityContext = securityContext;
-  }
-
   /**
    * Validates 'after' and 'before' parameters provided in search endpoints and creates an {@link IndexSelectCriteria}
-   * object based on those parameters. If the user does not have the 'threatIntelUseDailyIndices' permission the returned
-   * criteria will select the legacy 'act' index and the 'after' and 'before' parameters will be ignored. Otherwise, the
-   * returned criteria will select daily indices with 'indexStartTimestamp' and 'indexEndTimestamp' based on 'after' and 'before'.
+   * object based on those parameters. The returned criteria will select daily indices with 'indexStartTimestamp' and
+   * 'indexEndTimestamp' based on 'after' and 'before'.
    * <p>
    * If the 'before' timestamp is earlier than the 'after' timestamp an {@link InvalidArgumentException} will be thrown.
    * <p>
    * If the 'after' timestamp is earlier than the index retention of 1095 days an {@link InvalidArgumentException} will be thrown.
-   * <p>
-   * Validation of 'after' and 'before' is only performed if daily indices will be selected.
    *
    * @param after  'after' timestamp provided by search endpoints (optionally, defaults to before - 30 days)
    * @param before 'before' timestamp provided by search endpoints (optionally, defaults to now)
@@ -48,13 +33,6 @@ public class IndexSelectCriteriaResolver {
    * @throws InvalidArgumentException If the 'after' and 'before' parameters fail validation
    */
   public IndexSelectCriteria validateAndCreateCriteria(Long after, Long before) throws InvalidArgumentException {
-    if (!useDailyIndices()) {
-      // If the user isn't allowed/configured to use daily indices fall back to query the legacy 'act' index.
-      return IndexSelectCriteria.builder()
-              .setUseLegacyIndex(true)
-              .build();
-    }
-
     // Specify default values for end (now) and start (end - 30 days) if those values aren't provided.
     Instant end = ObjectUtils.ifNotNull(before, Instant::ofEpochMilli, clock.instant());
     Instant start = ObjectUtils.ifNotNull(after, Instant::ofEpochMilli, end.minus(30, ChronoUnit.DAYS));
@@ -73,19 +51,9 @@ public class IndexSelectCriteriaResolver {
     }
 
     return IndexSelectCriteria.builder()
-            .setUseLegacyIndex(false)
             .setIndexStartTimestamp(start.toEpochMilli())
             .setIndexEndTimestamp(end.toEpochMilli())
             .build();
-  }
-
-  private boolean useDailyIndices() {
-    try {
-      securityContext.checkPermission(TiFunctionConstants.threatIntelUseDailyIndices);
-      return true;
-    } catch (AccessDeniedException | AuthenticationFailedException ignored) {
-      return false;
-    }
   }
 
   /* Setters used for unit testing */