+++ service/testsrc/no/mnemonic/act/platform/service/ti/resolvers/IndexSelectCriteriaResolverTest.java
@@ -1,10 +1,7 @@
 package no.mnemonic.act.platform.service.ti.resolvers;
 
-import no.mnemonic.act.platform.api.exceptions.AccessDeniedException;
 import no.mnemonic.act.platform.api.exceptions.InvalidArgumentException;
 import no.mnemonic.act.platform.dao.api.criteria.IndexSelectCriteria;
-import no.mnemonic.act.platform.service.contexts.SecurityContext;
-import no.mnemonic.act.platform.service.ti.TiFunctionConstants;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -14,16 +11,15 @@ import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 
 import static no.mnemonic.act.platform.service.ti.resolvers.IndexSelectCriteriaResolver.MAXIMUM_INDEX_RETENTION_DAYS;
-import static org.junit.Assert.*;
-import static org.mockito.Mockito.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThrows;
+import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 public class IndexSelectCriteriaResolverTest {
 
   private static final Instant NOW = Instant.parse("2022-01-01T12:00:00Z");
 
-  @Mock
-  private SecurityContext securityContext;
   @Mock
   private Clock clock;
 
@@ -35,14 +31,7 @@ public class IndexSelectCriteriaResolverTest {
 
     when(clock.instant()).thenReturn(NOW);
 
-    resolver = new IndexSelectCriteriaResolver(securityContext).withClock(clock);
-  }
-
-  @Test
-  public void testSelectLegacyIndex() throws Exception {
-    doThrow(AccessDeniedException.class).when(securityContext).checkPermission(TiFunctionConstants.threatIntelUseDailyIndices);
-    assertTrue(resolver.validateAndCreateCriteria(null, null).isUseLegacyIndex());
-    verify(securityContext).checkPermission(TiFunctionConstants.threatIntelUseDailyIndices);
+    resolver = new IndexSelectCriteriaResolver().withClock(clock);
   }
 
   @Test
@@ -51,7 +40,6 @@ public class IndexSelectCriteriaResolverTest {
     long end = NOW.minus(1, ChronoUnit.DAYS).toEpochMilli();
 
     IndexSelectCriteria criteria = resolver.validateAndCreateCriteria(start, end);
-    assertFalse(criteria.isUseLegacyIndex());
     assertEquals(start, criteria.getIndexStartTimestamp());
     assertEquals(end, criteria.getIndexEndTimestamp());
   }
@@ -62,7 +50,6 @@ public class IndexSelectCriteriaResolverTest {
     long end = NOW.toEpochMilli();
 
     IndexSelectCriteria criteria = resolver.validateAndCreateCriteria(null, null);
-    assertFalse(criteria.isUseLegacyIndex());
     assertEquals(start, criteria.getIndexStartTimestamp());
     assertEquals(end, criteria.getIndexEndTimestamp());
   }
@@ -73,7 +60,6 @@ public class IndexSelectCriteriaResolverTest {
     long end = NOW.minus(2, ChronoUnit.DAYS).toEpochMilli();
 
     IndexSelectCriteria criteria = resolver.validateAndCreateCriteria(null, end);
-    assertFalse(criteria.isUseLegacyIndex());
     assertEquals(start, criteria.getIndexStartTimestamp());
     assertEquals(end, criteria.getIndexEndTimestamp());
   }
@@ -84,7 +70,6 @@ public class IndexSelectCriteriaResolverTest {
     long end = NOW.toEpochMilli();
 
     IndexSelectCriteria criteria = resolver.validateAndCreateCriteria(start, null);
-    assertFalse(criteria.isUseLegacyIndex());
     assertEquals(start, criteria.getIndexStartTimestamp());
     assertEquals(end, criteria.getIndexEndTimestamp());
   }