+++ CHANGELOG.md
@@ -66,6 +66,8 @@ For this use case one-legged Facts with bi-directional binding should be created
 ElasticSearch in order to avoid outdated names in the index. If an entity cannot be resolved a 412 response will be returned.
 
 ### Removed
+- (2022-10-21) Removed handling of the legacy 'act' index. Ensure that all installations have reindexed their data into
+daily indices, see [MIGRATIONS.md](MIGRATIONS.md).
 - (2022-02-11) Removed the old mechanism to check for already existing Facts, including the configuration property
 `act.fact.existence.check.use.cassandra` which is the new default (**breaking**). Facts added to installations before
 October 2021 might be duplicated as a result. *It is recommended to set up installations older than October 2021 from